update socket handler
